requirements.txt
beautifulsoup4==4.10.0
cloudinary==1.29.0
face_recognition==1.3.0
geopy==2.2.0
opencv_python==4.5.5.64
Pillow==9.0.1
psutil==5.9.0
pyttsx3==2.90
requests==2.27.1
